LayoutTests/fast/events/wheelevent-in-scrolling-div.html

<head>
    <script>
        var expectedScrollTop = 200;
        var expectedScrollLeft = 100;

        if (window.layoutTestController) {
            layoutTestController.dumpAsText();
            layoutTestController.waitUntilDone();
        }

        function dispatchWheelEvent()
        {
            var overflowElement = document.getElementById("overflow");
            if (overflowElement) {
                overflowElement.addEventListener("mousewheel", mousewheelHandler, false);
                var wheelEvent = document.createEvent("WheelEvent");
                if (wheelEvent) {
                    wheelEvent.initWebKitWheelEvent(-window.expectedScrollLeft, -window.expectedScrollTop, window, 0, 0, 0, 0, false, false, false, false);
                    overflowElement.dispatchEvent(wheelEvent);
                }
            }

            setTimeout('printScrollOffsets();', 100);
        }

        function passFailString(p)
        {
            var color = p?"green":"red";
            var text = p?"PASS":"FAIL";
            return "<span style='color:"+color+"'>"+text+"</span>";
        }

        function printScrollOffsets()
        {
            var consoleDiv = document.getElementById("offsetConsole");
            var overflowDiv = document.getElementById("overflow");
            if (consoleDiv && overflowDiv) {
                consoleDiv.innerHTML = "";

                consoleDiv.innerHTML += "scrollTop=" + overflowDiv.scrollTop + " (should be " + window.expectedScrollTop + "): ";
                consoleDiv.innerHTML += passFailString(overflowDiv.scrollTop == window.expectedScrollTop);
                consoleDiv.innerHTML += "<br>"

                consoleDiv.innerHTML += "scrollLeft=" + overflowDiv.scrollLeft + " (should be " + window.expectedScrollLeft + "): ";
                consoleDiv.innerHTML += passFailString(overflowDiv.scrollLeft == window.expectedScrollLeft);
                consoleDiv.innerHTML += "<br>"
            }

            if (window.layoutTestController)
                window.layoutTestController.notifyDone();
        }

        function mousewheelHandler(e)
        {
            var consoleDiv = document.getElementById("mousewheelConsole");
            if (consoleDiv) {
                consoleDiv.innerHTML = "";

                consoleDiv.innerHTML += "wheelDeltaY=" + e.wheelDeltaY + " (should be " + window.expectedScrollTop*-120 + "): ";
                consoleDiv.innerHTML += passFailString(e.wheelDeltaY == window.expectedScrollTop*-120);
                consoleDiv.innerHTML += "<br>"

                consoleDiv.innerHTML += "wheelDeltaX=" + e.wheelDeltaX + " (should be " + window.expectedScrollLeft*-120 + "): ";
                consoleDiv.innerHTML += passFailString(e.wheelDeltaX == window.expectedScrollLeft*-120);
                consoleDiv.innerHTML += "<br>"

                var expectedScroll = e.wheelDeltaY?window.expectedScrollTop:window.expectedScrollLeft;
                consoleDiv.innerHTML += "wheelDelta=" + e.wheelDelta + " (should be " + expectedScroll*-120 + "): ";
                consoleDiv.innerHTML += passFailString(e.wheelDelta == expectedScroll*-120);
                consoleDiv.innerHTML += "<br>"
            }
        }
    </script>
</head>

<body style="margin:0" onload="setTimeout('dispatchWheelEvent();', 100)">
    <div id="overflow" style="border:2px solid black;overflow:auto;height:200px;width:200px;">
        <div style="background-color:red;height:200px;width:400px;"></div>
        <div style="background-color:green;height:200px;width:400px;"></div>
        <div style="background-color:red;height:200px;width:400px;"></div>
    </div>
    <div id="offsetConsole"><span style="color:red">FAIL</span></div>
    <div id="mousewheelConsole"><span style="color:red">FAIL</span></div>
</body>
